feat(slate-core): add hasQuery and hasCommand
